m13_Django/pyproject.toml
[tool.poetry]
name = "m13_django"
version = "0.1.0"
description = ""
authors = ["Yuriy Kuchma <krabatua@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.4"
psycopg2 = "^2.9.5"
Pillow = "^9.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
